requirements.txt
deepspeech==0.9.3
future==0.18.3
iso8601==1.0.2
numpy==1.22
pkg-resources==0.0.0
pvporcupine==2.1.0
PyAudio==0.2.11
pygame==2.1.2
PyYAML==6.0
RPi.GPIO==0.7.0
serial==0.0.97
